Hoist ls-lint config to the repo root
Move the ls-lint config from `packages/fluent-vue/.ls-lint.yml` to a
root `.ls-lint.yml` so the convention applies repo-wide. Rename the
four kebab-case files in `unplugin-fluent-vue/src` to camelCase to
match the Vue/Vite ecosystem norm (and what `fluent-vue` already does).

- packages/unplugin-fluent-vue/src/directive-transform.ts → directiveTransform.ts
- packages/unplugin-fluent-vue/src/loader-query.ts → loaderQuery.ts
- packages/unplugin-fluent-vue/src/plugins/external-plugin.ts → plugins/externalPlugin.ts
- packages/unplugin-fluent-vue/src/plugins/sfc-plugin.ts → plugins/sfcPlugin.ts

Drop the per-package `ls-lint` script and devDep; move
`@ls-lint/ls-lint` to the root and run it once from there. Tests,
build output, and lint pass; only internal filenames change, no
public API impact.
