+++ package.json
@@ -8,7 +8,7 @@
     "test": "pnpm -r test",
     "typecheck": "pnpm -r typecheck",
     "lint": "eslint .",
-    "ls-lint": "pnpm -r ls-lint",
+    "ls-lint": "ls-lint",
     "prepare": "husky",
     "changeset": "changeset",
     "version": "changeset version",
@@ -18,6 +18,7 @@
     "@antfu/eslint-config": "^9.0.0",
     "@changesets/cli": "^2.31.0",
     "@changesets/get-github-info": "^0.8.0",
+    "@ls-lint/ls-lint": "^2.3.1",
     "eslint": "^10.4.0",
     "husky": "^9.1.7",
     "lint-staged": "^17.0.5"
