--- packages/fluent-vue/.ls-lint.yml
@@ -1,7 +1,0 @@
-ls:
-  '{src,__tests__}':
-    .js: kebab-case
-    .ts: camelCase | PascalCase
-    .d.ts: camelCase
-    .spec.ts: camelCase | PascalCase
-    .mock.ts: camelCase
