+++ packages/unplugin-fluent-vue/src/loaderQuery.ts
