+++ packages/unplugin-fluent-vue/src/rollup.ts
@@ -1,4 +1,4 @@
-export { rollupPlugin as ExternalFluentPlugin } from './plugins/external-plugin'
-export { rollupPlugin as SFCFluentPlugin } from './plugins/sfc-plugin'
+export { rollupPlugin as ExternalFluentPlugin } from './plugins/externalPlugin'
+export { rollupPlugin as SFCFluentPlugin } from './plugins/sfcPlugin'
 
 export type { ExternalPluginOptions, SFCPluginOptions } from './types'
