+++ packages/unplugin-fluent-vue/src/vite.ts
@@ -1,4 +1,4 @@
-export { vitePlugin as ExternalFluentPlugin } from './plugins/external-plugin'
-export { vitePlugin as SFCFluentPlugin } from './plugins/sfc-plugin'
+export { vitePlugin as ExternalFluentPlugin } from './plugins/externalPlugin'
+export { vitePlugin as SFCFluentPlugin } from './plugins/sfcPlugin'
 
 export type { ExternalPluginOptions, SFCPluginOptions } from './types'
