+++ packages/unplugin-fluent-vue/src/webpack.ts
@@ -1,4 +1,4 @@
-export { webpackPlugin as ExternalFluentPlugin } from './plugins/external-plugin'
-export { webpackPlugin as SFCFluentPlugin } from './plugins/sfc-plugin'
+export { webpackPlugin as ExternalFluentPlugin } from './plugins/externalPlugin'
+export { webpackPlugin as SFCFluentPlugin } from './plugins/sfcPlugin'
 
 export type { ExternalPluginOptions, SFCPluginOptions } from './types'
